refactor(types): port feed types etc to explicit imports
