+++ src/api/Feed.js
@@ -40,6 +40,7 @@ import type {
 } from '../common/types/tasks';
 import type { ElementsXhrError, ErrorResponseData } from '../common/types/api';
 import type { SelectorItems, SelectorItem } from '../common/types/core';
+import type { Comment, Comments, Tasks, Task, FeedItem, FeedItems, AppActivityItems } from '../common/types/feed';
 
 const TASK_NEW_INITIAL_STATUS = TASK_NEW_NOT_STARTED;
 const TASK = 'task';
