+++ src/common/types/core.js
@@ -1,6 +1,11 @@
 // @flow
 import { ITEM_TYPE_FOLDER, ITEM_TYPE_FILE, ITEM_TYPE_WEBLINK } from '../constants';
 
+type Order = {
+    by: SortBy,
+    direction: SortDirection,
+};
+
 type InlineNoticeType = 'warning' | 'error' | 'success' | 'info' | 'generic';
 
 type ItemType = typeof ITEM_TYPE_FOLDER | typeof ITEM_TYPE_FILE | typeof ITEM_TYPE_WEBLINK;
@@ -26,6 +31,26 @@ type UserMini = {
     type: 'user',
 };
 
+type User = {
+    avatar_url?: string,
+    email?: string,
+    id: string,
+    login?: string,
+    name: string,
+    type: 'user',
+};
+
+type UserCollection = {
+    entries?: Array<User>,
+    isLoaded?: boolean,
+    limit?: number,
+    next_marker?: string,
+    offset?: number,
+    order?: Array<Order>,
+    previous_marker?: string,
+    total_count?: number,
+};
+
 type ISODate = string;
 
 type MarkerPaginatedCollection<T> = {
@@ -44,11 +69,14 @@ type SelectorItem = {
 type SelectorItems = Array<SelectorItem>;
 
 export type {
+    Order,
     InlineNoticeType,
     ItemType,
     FileMini,
     FolderMini,
     UserMini,
+    User,
+    UserCollection,
     ISODate,
     MarkerPaginatedCollection,
     SelectorItem,
