+++ src/common/types/feed.js
@@ -1,7 +1,109 @@
 // @flow strict
+import type { User } from './core';
 
 // Feed item types that can receive deeplinks inline in the feed
 type FocusableFeedItemType = 'task' | 'comment';
 
-// eslint-disable-next-line import/prefer-default-export
-export type { FocusableFeedItemType };
+type BoxCommentPermission = {
+    can_delete?: boolean,
+    can_edit?: boolean,
+};
+
+type BoxTaskPermission = {
+    can_delete?: boolean,
+    can_update?: boolean,
+};
+
+// this is a subset of TaskNew, which imports as `any`
+type Task = {
+    created_at: string,
+    created_by: User,
+    id: string,
+    permissions: BoxTaskPermission,
+    type: 'task',
+};
+
+type Tasks = {
+    entries: Array<Task>,
+    next_marker: ?string,
+};
+
+type Comment = {
+    created_at: string,
+    created_by: User,
+    id: string,
+    is_reply_comment?: boolean,
+    message?: string,
+    modified_at: string,
+    permissions: BoxCommentPermission,
+    tagged_message: string,
+    type: 'comment',
+};
+
+type Comments = {
+    entries: Array<Comment>,
+    total_count: number,
+};
+
+type ActivityTemplateItem = {|
+    id: string,
+    type: 'activity_template',
+|};
+
+type AppItem = {|
+    icon_url: string,
+    id: string,
+    name: string,
+    type: 'app',
+|};
+
+type BaseAppActivityItem = {|
+    activity_template: ActivityTemplateItem,
+    app: AppItem,
+    created_by: User,
+    id: string,
+    rendered_text: string,
+    type: 'app_activity',
+|};
+
+type AppActivityAPIItem = {|
+    occurred_at: string,
+    ...BaseAppActivityItem,
+|};
+
+type AppActivityAPIItems = {
+    entries: Array<AppActivityAPIItem>,
+    total_count: number,
+};
+
+type AppActivityItem = {|
+    created_at: string,
+    permissions: BoxItemPermission,
+    ...BaseAppActivityItem,
+|};
+
+type AppActivityItems = {
+    entries: Array<AppActivityItem>,
+    total_count: number,
+};
+
+type FeedItem = Comment | Task | BoxItemVersion | AppActivityItem;
+
+type FeedItems = Array<FeedItem>;
+
+export type {
+    FocusableFeedItemType,
+    BoxCommentPermission,
+    Task,
+    Tasks,
+    Comment,
+    Comments,
+    ActivityTemplateItem,
+    AppItem,
+    AppActivityAPIItem,
+    AppActivityAPIItems,
+    AppActivityItem,
+    AppActivityItems,
+    FeedItem,
+    FeedItems,
+};
