+++ src/elements/content-sidebar/activity-feed/activity-feed/ActiveState.js
@@ -10,7 +10,7 @@ import Comment from '../comment';
 import TaskNew from '../task-new';
 import Version, { CollapsedVersion } from '../version';
 import withErrorHandling from '../../withErrorHandling';
-import type { FocusableFeedItemType } from '../../../../common/types/feed';
+import type { FocusableFeedItemType, FeedItem, FeedItems } from '../../../../common/types/feed';
 import type { SelectorItems } from '../../../../common/types/core';
 import type { GetAvatarUrlCallback, GetProfileUrlCallback } from '../../../common/flowTypes';
 import type { Translations } from '../../flowTypes';
