+++ src/elements/content-sidebar/activity-feed/activity-feed/ActivityFeed.js
@@ -16,7 +16,7 @@ import { collapseFeedState, ItemTypes } from './activityFeedUtils';
 import InlineError from '../../../../components/inline-error/InlineError';
 import LoadingIndicator from '../../../../components/loading-indicator/LoadingIndicator';
 import messages from './messages';
-import type { FocusableFeedItemType } from '../../../../common/types/feed';
+import type { FocusableFeedItemType, FeedItems } from '../../../../common/types/feed';
 import type { SelectorItems } from '../../../../common/types/core';
 import type { GetAvatarUrlCallback, GetProfileUrlCallback } from '../../../common/flowTypes';
 import type { Translations } from '../../flowTypes';
