+++ src/elements/content-sidebar/activity-feed/activity-feed/activityFeedUtils.js
@@ -5,6 +5,7 @@
 
 import selectors from '../../../common/selectors/version';
 import { PLACEHOLDER_USER, VERSION_UPLOAD_ACTION } from '../../../../constants';
+import type { FeedItems } from '../../../../common/types/feed';
 
 export const ItemTypes = {
     fileVersion: 'file_version',
