+++ src/elements/content-sidebar/activity-feed/comment/Comment.js
@@ -22,6 +22,7 @@ import messages from './messages';
 import type { GetAvatarUrlCallback, GetProfileUrlCallback } from '../../../common/flowTypes';
 import type { Translations } from '../../flowTypes';
 import type { SelectorItems } from '../../../../common/types/core';
+import type { BoxCommentPermission } from '../../../../common/types/feed';
 import './Comment.scss';
 
 type Props = {
